requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
joblib==0.13.2
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.3.3
matplotlib==3.1.1
networkx==2.5
numpy==1.17.0
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==0.25.0
patsy==0.5.1
pkg-resources==0.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.21.3
scikit-surprise==1.0.6
scipy==1.3.0
six==1.12.0
statsmodels==0.10.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
